Refactor listen/dispatch events from map
